fix the model and build issues && update README
