+++ app/model/unet_parts.py
@@ -1,11 +1,11 @@
-""" Components of the Watermark Removal U-Net Model """
+""" Parts of the U-Net model """
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
 
-class DualConvLayer(nn.Module):
+class DoubleConv(nn.Module):
     """(convolution => [BN] => ReLU) * 2"""
 
     def __init__(self, in_channels, out_channels):
@@ -23,49 +23,49 @@ class DualConvLayer(nn.Module):
         return self.double_conv(x)
 
 
-class DownscaleLayer(nn.Module):
-    """DownscaleLayerscaling with maxpool then double conv"""
+class Down(nn.Module):
+    """Downscaling with maxpool then double conv"""
 
     def __init__(self, in_channels, out_channels):
         super().__init__()
         self.maxpool_conv = nn.Sequential(
             nn.MaxPool2d(2),
-            DualConvLayer(in_channels, out_channels)
+            DoubleConv(in_channels, out_channels)
         )
 
     def forward(self, x):
         return self.maxpool_conv(x)
 
 
-class UpscaleLayer(nn.Module):
-    """UpscaleLayerscaling then double conv"""
+class Up(nn.Module):
+    """Upscaling then double conv"""
 
     def __init__(self, in_channels, out_channels, bilinear=True):
         super().__init__()
 
         # if bilinear, use the normal convolutions to reduce the number of channels
         if bilinear:
-            self.up = nn.UpscaleLayersample(scale_factor=2, mode='bilinear', align_corners=True)
+            self.up = nn.Upsample(scale_factor=2, mode='bilinear', align_corners=True)
         else:
             self.up = nn.ConvTranspose2d(in_channels // 2, in_channels // 2, kernel_size=2, stride=2)
 
-        self.conv = DualConvLayer(in_channels, out_channels)
+        self.conv = DoubleConv(in_channels, out_channels)
 
     def forward(self, x1, x2):
         x1 = self.up(x1)
-        # input is CHW
         diffY = torch.tensor([x2.size()[2] - x1.size()[2]])
         diffX = torch.tensor([x2.size()[3] - x1.size()[3]])
 
         x1 = F.pad(x1, [diffX // 2, diffX - diffX // 2,
                         diffY // 2, diffY - diffY // 2])
+
         x = torch.cat([x2, x1], dim=1)
         return self.conv(x)
 
 
-class OutputConvLayer(nn.Module):
+class OutConv(nn.Module):
     def __init__(self, in_channels, out_channels):
-        super(OutputConvLayer, self).__init__()
+        super(OutConv, self).__init__()
         self.conv = nn.Conv2d(in_channels, out_channels, kernel_size=1)
 
     def forward(self, x):
