+++ app/model/vgg.py
@@ -1,4 +1,3 @@
-import torch
 import torch.nn as nn
 from torchvision import models
